Don't wait for tests to pass before starting all platform builds
Some platform builds take a while to get warmed up and although
obviously you wouldn't want to build things that are bad (i.e. failing
tests) making them wait for the tests to pass before even starting is
making the build process unecessarily long, especially considering that
the project doesn't even have any tests to start with.

Instead run the builds for all platforms in parallel *with* the tests
and instead make the last "bundle upload" step only depend on the tests,
so maybe some builds will fail or the tests will fail and then in either
case nothing will be uploaded.  The result should be similar to the
current flow in usefulness but with wasting much less time.