Load and draw the game map tiles from LDTk
The renderer is the same embedFS-based renderer as in previous projects,
I just copied it across (this should really be put in a lib).  The tiles
and background colour are loaded from the LDTk project and pre-rendered
into an image.  This image is now also drawn onto the screen during
Draw, it's in the wrong place but at least it works and it's progress.