Remove slashes from tileset references in LDTK project
This caused an issue where the game wouldn't start up on Windows because
the ldtk library we're using flips the / to \ for Windows and can't find
the file inside embedfs where all the slashes are forward slashes.
Irritating but easily fixed by just dropping the ./ from the start of
the relative path.