Add tiles to collision space based on their ID
This introduces 4 tile types:
- climbable
- wall
- chasm
- slippery

These are added to the resolv space by iterating over the ldtk layer and
checking the ID of the tile, each ID of a tile image corresponds to a
type of tile.

Now you will only collide with wall tiles.  For the rest, its type will
be printed out to the screen as debug info and their behaviour will be
implemented later.