Add collision debugger
It works but draws the boxes in the wrong place...

This toggleable/plugable debugger system is re-used from last year, I
should turn it into a lib soon.  The collision debugger draws squares
around all the objects in the resolv space so you can easily visualise
why things are (or are NOT!) colliding.  To use it run the program with
the build tag:

    go run -tags debugcol .