Convert JSON files' line endings from DOS to UNIX
The JSON files are exported from Aseprite on a Windows computer so they
have DOS line endings and copying them to the repo shows the whole file
as having been modified just because of the line endings (second time
this happens now).  To avoid forgetting to convert it in future this git
attributes setting will convert them automatically when adding the file.