Set resolv space to the map size instead of the screen size 🤦
Map tiles below the bottom of the screen weren't showing any effect on
the debug output and then I realised the mistake:  I'd used gameHeight
and gameWidth for initialising the resolv Space but that's just the
logical size of the viewport/camera.

I've fixed it to use the current map level's size now and it works as
expected.  This also means the player initialisation had to be moved
down lower because it needs the space to be initialised first, but thad
had to come after the map rendering because the map level doesn't exist
before that.