Fix sprite-object alignment by centering the player sprite
There was an issue where collision seemed to work from two directions,
but allowed you to walk halfway into the tiles from the other two
directions.  This was because of a mismatch of position of where the
objects were located in the collision space and where they were being
drawn on the screen.

Assuming the tiles are all rendered correctly I removed the
"recentering" to align their objects with their pre-rendered drawing.
Then I did the same for the player object and noticed that to center it,
I need to adjust where the sprite is drawn relative to the object,
instead of the other way around.  The collision box is 8x8, half the
size of the sprite, so to center the sprite on it, we translate the
sprite negatively by a 1/4 of its size.  Now everything is perfectly
lined up and the collisions match what you see on the screen.