Implement falling mechanic
If, after jumping, you land on a tile that's a chasm, instead of going
to idle state go to falling.  Currently the wall-collision logic from
climbing is re-used, so when you hit a wall the velocity goes to zero,
and in the case of falling it'll switch to falling recovery animation.

I'm sure there's some edge cases I've missed but I'll get back to it.

The different states and transitions and conditions for transitions are
getting more and more complicated, it's getting more urgent to take a
stricter FSM-based approach, I think.