Remove Q hotkey for quitting the game
This has been really annoying during play testing because it's right
next to the movement controls, so it's really easy to accidentally press
Q and quit the game.  I haven't replaced it with anything because the
mouse is available so you can close the game by pressing the [X] button
the game window decoration.