Behave like intersectoin when one obj is completely inside the other
It turns out the collision respsonse library only considers something as
to be intersecting if their edges are being crossed.  Usually this isn't
an issue because that'll happen when you walk into it. OTOH at the end
of the jump if you land perfectly inside a tile you mnay miss the edge
of it as you land and will only start falling once you mve enough to
keep the balance on metre.  Now with insideOF we can also check if a
shape is fully contained by another... or at least rectangular shapes.