Display current animation state on screen for debugging
So far these were only referenced internally as integer constants to
avoid typos but we'd like to display their string values so it's easier
to follow without looking up which state is #2 or #3.  The sprite tags
generation tool has been expanded to include an array of tag names as
strings, which can be looked up using the integer constants that are in
use in the game already.