Update all Nanobot sprite animations and tags
The update sprite animations have a simplified and more consistent
naming structure which also means the tag names had to be updated
everywhere in the code.  They also have separate animation frames for
moving in all 4 directions, so the movement code has been updated to use
these instead of the previous horizontal and vertical animations.  This
is because the original idea of flipping the sprite in code won't work
because the robot isn't symmetrical.