Improve jump end animations
When the bot lands cleanly on a climbable tile use the "jump land floor"
animation, the other one is for when it bumps a wall.  The animation
looked a bit moonwalky so I added some small speed changes during the
start and end animations: for start slightly backwards, charging up for
the jump, for end, slightly forwards as it pulls itself on the tile.