Move Falling/Slipping state switch to after the animation completes
There was an issue where you'd be falling and as you reach the end of
the chasm/slippery it'd turn off the state and therefore skip the end
fall/slip animation and mess up the state changes for a while.  That's
because, like with jumping, the state change needs to happen _after_ the
end animation has completed, so I moved them to there and it's better.