Only recover onto tiles below you
Pass down the intersection object from the outer check so that we can
get the Minimum Translation Vector (MTV) that would be needed to move to
outside of the object.  Not because we want to move out of it (it is
climbable, we are intentionally falling down onto it) but because
whether it's negative or positive will tell us whether we're falling
down onto it from above or not.  This is important because there was an
issue where during falling you might pass by some more climbable tiles
on the way down, close enough to be intersecting a bit maybe, but it's
not the one below you.  To only stop and land on tiles below you that
means the MTV to get out of them will be negative, i.e. upwards, because
the origin (0, 0) is at the top left.