Switch jump max from airtime to distance based
Leaning on the Vector package I saw being used by resolv I check how far
the player has travelled during jumping and then stop it once you've
reached the max.  Something is still not quite right with it though
because even after setting the minimum distance to zero it seems to keep
jumping, as if the player is also moving during the jump start tag.