Add a green glowing light effect under the player
I just used a white PNG of an airbrush effect, added transparency and
green colour and placed it under the player.  Because the player is
drawn a layer below the walls it also has the cool effect that the light
doesn't go through the walls, without having to do any extra work.