Update nanobot sprite and animations
This includes renaming for some states from start/end to separate start
and end tags and also moves the sprite on the sprite sheet during jump
animation so that it doesn't have to be moved in code, which causes
annoying issues with the game physics.  These code compensations will
have to be removed next.