Override internal Nanobot sprite from OS if found
To be able to iterate faster on the animations it's nice to be able to
use a local Nanobot sprite without having to rebuild the game.  If it
finds Nanobot.png and Nanobot.png in the same folder as the game it'll
try to use those instead.  As a fallback the internal asset is always
used, also in case of failure.  This feature should be turned off for
the production build.