Add concept of scenes and a start scene that shows before the game
I wanted to use stagehand to manage the scenes but I didn't like having
to implement Layout() for compliance with ebiten.Game.  Probably just
should've bitten the bullet because I ended up having to adjust the
debuggers accordingly, which was pretty annoying.

Anyway now there's a start scene and an implemented mechanism for
switching between scenes.