Add continuously rising water that kills you if you fall in it
This included adding a death screen and winning screen instead of the
previous approach of quitting the game with an error.  The water is a
new entity that is continuously rising and there's a check that if the
water is higher than the top of your sprite then you die.  This is also
cool because you don't need a second check for death from falling
because after falling for a while you will fall into the water anyway.