Apply low-pass filter to music upon death
This is to make the music sound like it's playing underwater, because
you die when you fall into the water.  The resound library is used for
the low-pass filter because DSPs are hard.  The DSP needs to applied
in-between when OGG data is read and an audio.Player is made out of it,
so I merged the NewSoundPlayer function code directly into AddSound
because anyway this was the only place it was being called from.