Increase AnimationSkipTicks from 5 to 6
6 makes for nicer, rounder numbers when trying to sync with other things
like sound effects.  I realise now that I probably named this constant
badly, thinking it would be how many frames it skips, but it itself is
included in the last tick, so maybe a better name would hint that it is
"holding" the animation from for so many ticks.  Anyway...