Add flickering effect for the player's light
This is just to be fancy and to give more the effect that the glow in
front of you comes from a light-source, it's flickering like an old
fluorescent tube or some other light source with inconsistent power
supply.  Perlin noise is used for the flickering variation to make the
randomness feel more natural.